Try with a lock
